security improvements

+++ src/main/java/hu/unideb/notetakingapp/frontend/controller/NotesController.java
@@ -43,21 +43,24 @@ public class NotesController {
     }
 
     @GetMapping("/notes")
-    public String LoginForm(Model model) {
-        if (loggedInUserBean.isLoggedIn()) {
-            model.addAttribute("new_note", new Note());
-            model.addAttribute("user", loggedInUserBean.getLoggedInUser());
-            model.addAttribute("notes", noteService.findAll().stream()
-                    .filter(note -> note.getCreatorUser().getUserName().equals(loggedInUserBean.getLoggedInUser().getUserName()))
-                    .sorted(Comparator.comparing(Note::getTitle))
-                    .toArray());
-            return "notes";
-        } else
+    public String Notes(Model model) {
+        if (!loggedInUserBean.isLoggedIn())
             return "redirect:/login";
+
+        model.addAttribute("new_note", new Note());
+        model.addAttribute("user", loggedInUserBean.getLoggedInUser());
+        model.addAttribute("notes", noteService.findAll().stream()
+                .filter(note -> note.getCreatorUser().getUserName().equals(loggedInUserBean.getLoggedInUser().getUserName()))
+                .sorted(Comparator.comparing(Note::getTitle))
+                .toArray());
+        return "notes";
     }
 
     @PostMapping("/noteadd")
     public String addNote(@ModelAttribute Note note, Model model) {
+        if (!loggedInUserBean.isLoggedIn())
+            return "redirect:/login";
+
         note.setCreationDate(LocalDate.now());
         User actUser = userService.findByUsername(loggedInUserBean.getLoggedInUser().getUserName());
         note.setCreatorUser(actUser);
@@ -67,6 +70,9 @@ public class NotesController {
 
     @GetMapping("/noteDelete")
     public String deleteNote(@RequestParam Long id, Model model) {
+        if (!loggedInUserBean.isLoggedIn())
+            return "redirect:/login";
+
         Note toDelete = noteService.findById(id);
         noteService.delete(toDelete);
         return "redirect:/notes";
@@ -74,6 +80,9 @@ public class NotesController {
 
     @GetMapping("/initUpdate")
     public String initUpdate(@RequestParam Long id, Model model) {
+        if (!loggedInUserBean.isLoggedIn())
+            return "redirect:/login";
+
         Note updatable = noteService.findById(id);
         selectedNoteBean.setSelectedNote(updatable);
         isFormVisibleBean.setVisible(true);
@@ -82,6 +91,9 @@ public class NotesController {
 
     @PostMapping("/updateNote")
     public String updateNote(@ModelAttribute("note") Note note, Model model) {
+        if (!loggedInUserBean.isLoggedIn())
+            return "redirect:/login";
+
         selectedNote().setTitle(note.getTitle());
         selectedNote().setBody(note.getBody());
         selectedNote().setLastModifiedDate(LocalDate.now());
